Start migration of onObjectCreated to ingestion manager
