+++ services/ingestion/manager/.babelrc
@@ -11,6 +11,7 @@
       }
     ]
   ],
+  "plugins": ["transform-object-rest-spread"],
   "env": {
     "test": {
       "presets": [
