+++ services/ingestion/manager/package.json
@@ -7,7 +7,10 @@
     "test:unit": "jest --testPathPattern=unit"
   },
   "dependencies": {
-    "@eubfr/logger-listener": "0.0.1"
+    "@eubfr/logger-listener": "0.0.1",
+    "babel-plugin-transform-object-rest-spread": "6.26.0",
+    "elasticsearch": "14.0.0",
+    "http-aws-es": "4.0.0"
   },
   "devDependencies": {
     "aws-sdk": "2.179.0",
