+++ services/storage/meta-index/serverless.yml
@@ -35,21 +35,6 @@ provider:
       Resource: "arn:aws:es:${self:provider.region}:*:domain/*"
 
 functions:
-  onObjectCreated:
-    handler: src/events/onObjectCreated.handler
-    name: ${self:provider.stage}-${self:service}-onObjectCreated
-    events:
-      - sns:
-          arn:
-            Fn::Join:
-              - ""
-              -
-                - "arn:aws:sns:"
-                - Ref: "AWS::Region"
-                - ":"
-                - Ref: "AWS::AccountId"
-                - ":${self:provider.stage}-object-created"
-          topicName: ${self:provider.stage}-object-created
   onObjectRemoved:
     handler: src/events/onObjectRemoved.handler
     name: ${self:provider.stage}-${self:service}-onObjectRemoved
